requirements.txt
cloudinary==1.39.1
channels==4.1.0
daphne==4.1.0
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.2
django-allauth==0.63.1
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
Jinja2==3.1.2
mysqlclient==2.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.8.0
openapi-codec==1.3.2
pandas==2.2.0
psycopg2==2.9.9
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
resend==1.1.0
SQLAlchemy==2.0.25
virtualenv==20.25.0
